added new routes reports
