+++ src/components/InfoPage/InfoPage.js
@@ -15,7 +15,7 @@ class InfoPage extends Component {
 
   componentDidUpdate() {
     if (!this.props.user.isLoading && this.props.user.userName === null) {
-      this.props.history.push('home');
+      this.props.history.push('login');
     }
   }
